+++ pg_lakehouse/src/api/mod.rs
@@ -7,7 +7,7 @@ use crate::datafusion::context::ContextError;
 use crate::datafusion::format::*;
 use crate::datafusion::provider::*;
 use crate::datafusion::session::Session;
-use crate::stores::utils::*;
+use crate::fdw::utils::*;
 
 #[pg_extern]
 pub fn arrow_schema(
