+++ pg_lakehouse/src/fdw/mod.rs
@@ -1,2 +1,6 @@
+pub mod base;
 pub mod handler;
+pub mod local;
 pub mod options;
+pub mod s3;
+pub mod utils;
