+++ pg_lakehouse/src/fdw/utils.rs
