+++ pg_lakehouse/src/lib.rs
@@ -3,7 +3,6 @@ mod datafusion;
 mod fdw;
 mod hooks;
 mod schema;
-mod stores;
 
 use hooks::LakehouseHook;
 use pgrx::*;
